moved middleware to its own folder
